Added new features and improved code structure
This commit introduces several new features and improvements to the codebase. A `clipboard` option was added to the `App` class, and the `get_config` method was updated to check for each option in the `updated_dictionary`. Duplicate imports from the `code` string are now removed. Lists and dictionaries were added to specify and map the modules to be included when building the executable file with Nuitka and PyInstaller. The error message in the `write_and_obfuscate` method was updated to be printed to the console in addition to being logged.

Significant changes were made to the luna.py file, including the removal of several import statements, modification of how threads are added to the `threads` list, and replacement of the `zipup()` function call with a block of code that performs the same task. The `temp_path` variable is now a global variable.

New classes were added in `AntiSpam.py`, `BackupCodes.py`, `Browsers.py`, and `Clipboard.py` to perform various tasks such as checking for file modifications, searching for files, killing browser processes, and writing clipboard content to a file.

Changes were made in `Debug.py` and `disable_defender.py` to perform system checks and disable Windows Defender features. New files were added for extracting Discord tokens, displaying fake error messages, injecting JavaScript code into Discord's core files, disabling the Discord Token Protector, extracting Minecraft account data, extracting system information, deleting the script itself, ensuring the script runs on system startup, and extracting saved Wi-Fi network profiles.
